+++ src/hash_util.c
@@ -40,7 +40,7 @@ PRIVATE void          *hashtab[HASHSIZE + 1];
 
 PUBLIC unsigned long  collisions = 0;
 
-static unsigned long  hashfillmax = HASHSIZE * 2. / 3.;
+static unsigned long  hashfillmax = (unsigned long)(HASHSIZE * 2. / 3.);
 //static unsigned long  hashfillmax = HASHSIZE;
 
 unsigned long         Number_of_hash_entries = 0;
